leveldb: re-organize
